Fixed encoder and decoder for 32-bit arch

+++ source/styx2000/protomsg/decoder.d
@@ -29,7 +29,7 @@ private {
 		E listOfTypes;
 		StyxObject[] msg;
 		
-		ulong vlsPosition = 0;
+		size_t vlsPosition = 0;
 	
 		foreach (e; listOfTypes)
 		{
